+++ src/tree/traits.rs
@@ -7,6 +7,7 @@
 //!   - [custom_tree_vec](https://github.com/DioxusLabs/taffy/blob/main/examples/custom_tree_vec.rs) which implements a custom Taffy tree using a `Vec` as an arena with NodeId's being index's into the Vec.
 //!   - [custom_tree_owned_partial](https://github.com/DioxusLabs/taffy/blob/main/examples/custom_tree_owned_partial.rs) which implements a custom Taffy tree using directly owned children with NodeId's being index's into vec on parent node.
 //!   - [custom_tree_owned_unsafe](https://github.com/DioxusLabs/taffy/blob/main/examples/custom_tree_owned_unsafe.rs) which implements a custom Taffy tree using directly owned children with NodeId's being pointers.
+//!   - [custom_tree_owned_unsafe_linked](https://github.com/DioxusLabs/taffy/blob/main/examples/custom_tree_owned_unsafe_linked.rs) same to `custom_tree_owned_unsafe` but use linked list to storage childrens.
 //!
 //! ## Overview
 //!
@@ -43,20 +44,15 @@
 //! # use taffy::*;
 //! pub trait TraversePartialTree {
 //!     /// Type representing an iterator of the children of a node
-//!     type ChildIter<'a>: Iterator<Item = NodeId>
-//!     where
-//!         Self: 'a;
+//!     type ChildIter: SuspendIterator<Self, Item = NodeId>;
 //!
 //!     /// Get the list of children IDs for the given node
-//!     fn child_ids(&self, parent_node_id: NodeId) -> Self::ChildIter<'_>;
+//!     fn child_ids(&self, parent_node_id: NodeId) -> Self::ChildIter;
 //!
 //!     /// Get the number of children for the given node
 //!     fn child_count(&self, parent_node_id: NodeId) -> usize;
-//!
-//!     /// Get a specific child of a node, where the index represents the nth child
-//!     fn get_child_id(&self, parent_node_id: NodeId, child_index: usize) -> NodeId;
 //! }
-//!
+//! 
 //! pub trait TraverseTree: TraversePartialTree {}
 //! ```
 //!
@@ -142,23 +138,77 @@ use crate::{BlockContainerStyle, BlockContext, BlockItemStyle};
 #[cfg(all(feature = "grid", feature = "detailed_layout_info"))]
 use crate::compute::grid::DetailedGridInfo;
 
+/// Suspendable iterator.
+/// It extract the state required for iteration,
+/// without holding a persistent borrow of `Source`. 
+/// Instead, borrow `Source` only temporarily during each call to `next`.
+pub trait SuspendIterator<Source: ?Sized> {
+    /// The type of the elements being iterated over.
+    type Item;
+
+    /// Advances the iterator and returns the next value. 
+    /// Also returns the current iteration count.
+    fn next(&mut self, source: &Source) -> Option<(usize, Self::Item)>;
+
+    /// Convert to an iterator for ease of use.
+    fn iter(self, source: &'_ Source) -> SuspendIter<'_, Self, Source>
+    where
+        Self: Sized,
+    {
+        SuspendIter(self, source)
+    }
+
+    /// Convert to an enumerate iterator for ease of use.
+    fn iter_enumerate(self, source: &'_ Source) -> SuspendIterWithIndex<'_, Self, Source>
+    where
+        Self: Sized,
+    {
+        SuspendIterWithIndex(self, source)
+    }
+}
+
+/// Immutable iterator wrapper for ease of use.
+#[doc(hidden)]
+pub struct SuspendIter<'a, Suspend, Source: ?Sized>(Suspend, &'a Source);
+
+impl<'a, Suspend, Source: ?Sized> Iterator for SuspendIter<'a, Suspend, Source>
+where
+    Suspend: SuspendIterator<Source>,
+{
+    type Item = Suspend::Item;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.0.next(self.1).map(|a| a.1)
+    }
+}
+
+/// Mutable iterator wrapper for ease of use.
+#[doc(hidden)]
+pub struct SuspendIterWithIndex<'a, Suspend, Source: ?Sized>(Suspend, &'a Source);
+
+impl<'a, Suspend, Source: ?Sized> Iterator for SuspendIterWithIndex<'a, Suspend, Source>
+where
+    Suspend: SuspendIterator<Source>,
+{
+    type Item = (usize, Suspend::Item);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.0.next(self.1)
+    }
+}
+
 /// Taffy's abstraction for downward tree traversal.
 ///
 /// However, this trait does *not* require access to any node's other than a single container node's immediate children unless you also intend to implement `TraverseTree`.
 pub trait TraversePartialTree {
     /// Type representing an iterator of the children of a node
-    type ChildIter<'a>: Iterator<Item = NodeId>
-    where
-        Self: 'a;
+    type ChildIter: SuspendIterator<Self, Item = NodeId>;
 
     /// Get the list of children IDs for the given node
-    fn child_ids(&self, parent_node_id: NodeId) -> Self::ChildIter<'_>;
+    fn child_ids(&self, parent_node_id: NodeId) -> Self::ChildIter;
 
     /// Get the number of children for the given node
     fn child_count(&self, parent_node_id: NodeId) -> usize;
-
-    /// Get a specific child of a node, where the index represents the nth child
-    fn get_child_id(&self, parent_node_id: NodeId, child_index: usize) -> NodeId;
 }
 
 /// A marker trait which extends `TraversePartialTree`